# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface QuizRequest {
  /** CBSE, ICSE, SCERT, or Mixed */
  board: string;
  /** Class 6 through Class 12 */
  className: string;
  subject: string;
  chapter: string;
  /** Easy, Medium, Hard, or Mixed */
  level: string;
  questionCount?: number;
}

export interface ChapterRequest {
  board: string;
  className: string;
  subject: string;
  chapter: string;
}

export interface QuizQuestion {
  id: number;
  sessionId?: number;
  questionText: string;
  options: string[];
  correctAnswer: string;
  explanation: string;
  /** previous_year or ai_generated */
  type: string;
  /** @nullable */
  year?: string | null;
  topic?: string;
  difficulty?: string;
}

export interface QuizSession {
  id: number;
  board: string;
  className: string;
  subject: string;
  chapter: string;
  level: string;
  /** in_progress or completed */
  status: string;
  /** @nullable */
  score?: number | null;
  totalQuestions?: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  questions: QuizQuestion[];
}

export interface QuizSessionSummary {
  id: number;
  board: string;
  className: string;
  subject: string;
  chapter: string;
  level: string;
  status: string;
  /** @nullable */
  score?: number | null;
  totalQuestions: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface QuizCompletion {
  score: number;
}

export interface TopicFrequency {
  topic: string;
  count: number;
  subject: string;
  chapter: string;
}

export interface SubjectBreakdown {
  subject: string;
  count: number;
  averageScore: number;
}

export interface QuizStats {
  totalSessions: number;
  totalQuestions: number;
  averageScore: number;
  topicFrequency: TopicFrequency[];
  subjectBreakdown: SubjectBreakdown[];
}

export interface FormulaFormula {
  name: string;
  formula: string;
  description: string;
  unit: string;
  /** @nullable */
  derivation?: string | null;
}

export interface FormulaSection {
  sectionTitle: string;
  formulas: FormulaFormula[];
}

export interface FormulaSheet {
  subject: string;
  chapter: string;
  sections: FormulaSection[];
  previousYearTopics: string[];
}

export interface ShortNotesItem {
  heading: string;
  content: string;
  keyPoints: string[];
  equations?: string[];
  /** @nullable */
  previousYearRelevance?: string | null;
}

export interface PreviousYearDataPoint {
  year: string;
  questionsCount: number;
  topics: string[];
}

export interface ShortNotes {
  subject: string;
  chapter: string;
  items: ShortNotesItem[];
  previousYearGraph: PreviousYearDataPoint[];
}

export interface DocumentUpload {
  fileName: string;
  /** Base64-encoded file content */
  fileData: string;
  mimeType: string;
}

export interface DocumentAnalysis {
  extractedText: string;
  /** @nullable */
  detectedSubject: string | null;
  /** @nullable */
  detectedChapter: string | null;
  suggestedQuestions: string[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}