# lib/api-zod/src/generated/types/openaiVoiceMessageInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}